broke bots.tsx into smaller components and renamed / refactored some junk in the AI serverless function
