+++ src/components/bot_management/index.ts
@@ -1,0 +1,5 @@
+export { default as BotTable } from './bot-table'
+export { default as AddBotDialog } from './add-bot-dialog'
+export { default as EditBotDialog } from './edit-bot-dialog'
+export { default as TestBotDialog } from './test-bot-dialog'
+export { default as TestResultDialog } from './test-result-dialog'
